// SistemaInventarios/src/app/Interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
    type:        string;
    query:       string[];
    features:    Feature[];
    attribution: string;
}

export interface Feature {
    id:            string;
    type:          string;
    place_type:    string[];
    relevance:     number;
    properties:    Properties;
    text_es:       string;
    place_name_es: string;
    text:          string;
    place_name:    string;
    center:        number[];
    geometry:      Geometry;
    address:       string;
    context:       Context[];
}

export interface Context {
    id:           string;
    wikidata?:    Wikidata;
    text_es:      Text;
    language_es?: Language;
    text:         Text;
    language?:    Language;
    short_code?:  ShortCode;
}

export enum Language {
    Es = "es",
}

export enum ShortCode {
    Gt = "gt",
    GtGu = "GT-GU",
}

export enum Text {
    CiudadDeGuatemala = "Ciudad de Guatemala",
    ColoniaLaForesta = "Colonia La Foresta",
    Guatemala = "Guatemala",
}

export enum Wikidata {
    Q1555 = "Q1555",
    Q695660 = "Q695660",
    Q774 = "Q774",
}

export interface Geometry {
    type:          string;
    coordinates:   number[];
    interpolated?: boolean;
    omitted?:      boolean;
}

export interface Properties {
    accuracy:             string;
    "override:postcode"?: string;
}